HHH-3646 - don't append spurious comma in SQL select fragment
In the JoinWalker, we assemble an SQL select list fragment
from a list of Joinables. In the existing code, we see it's
possible for the selectFragment generated by a joinable to be empty.
However if the LAST joinable generates an empty fragment, we put
a spurious comma into the fragment. Fix this by only prepending a
comma when necessary.